refactor: job과 occupation 하나의 패키지로 정리
